web_scraping_project/requirements.txt
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
DateTime==6.0
google-api-core==2.28.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gspread==6.2.1
gspread-dataframe==4.0.0
httplib2==0.31.0
idna==3.11
lxml==6.0.2
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.3.1
pandas==2.3.3
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.8.1
time-delta==0.0.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
zope.interface==8.1.1
